:recycle: REFACTOR: statistic service 1차 통합

+++ slash/src/main/java/project/slash/statistics/dto/GradeScoreDto.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public class GradeScoreDto {
 	private String grade;
-	private double score;
+	private int score;
 
 }
